refactor(shorebird_cli): remove old release commands
