+++ packages/shorebird_cli/lib/src/command_runner.dart
@@ -7,7 +7,7 @@ import 'package:mason_logger/mason_logger.dart';
 import 'package:scoped/scoped.dart';
 import 'package:shorebird_cli/src/commands/commands.dart';
 import 'package:shorebird_cli/src/commands/patch_new/patch_new.dart';
-import 'package:shorebird_cli/src/commands/release_new/release_new.dart';
+import 'package:shorebird_cli/src/commands/release/release.dart';
 import 'package:shorebird_cli/src/engine_config.dart';
 import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/platform.dart';
@@ -78,7 +78,6 @@ class ShorebirdCliCommandRunner extends CompletionCommandRunner<int> {
     addCommand(PatchNewCommand());
     addCommand(PreviewCommand());
     addCommand(ReleaseCommand());
-    addCommand(ReleaseNewCommand());
     addCommand(RunCommand());
     addCommand(UpgradeCommand());
   }
