--- packages/shorebird_cli/lib/src/commands/release_new/release_new.dart
@@ -1,6 +1,0 @@
-export 'aar_releaser.dart';
-export 'android_releaser.dart';
-export 'ios_framework_releaser.dart';
-export 'ios_releaser.dart';
-export 'release_new_command.dart';
-export 'releaser.dart';
